+++ mosqito/functions/oct3filter/comp_third_spectrum.py
@@ -9,9 +9,6 @@ Created on Mon Nov 16 09:10:45 2020
 import numpy as np
 
 # Local imports
-from mosqito.functions.oct3filter.calc_third_octave_levels import (
-    calc_third_octave_levels,
-)
 from mosqito.functions.oct3filter.oct3spec import oct3spec
 
 
@@ -40,8 +37,8 @@ def comp_third_spec(is_stationary, signal, fs):
     if is_stationary == True:
         spec_third, third_axis = oct3spec(signal, fs)
         time_axis = []
-    elif is_stationary == False:
-        spec_third, third_axis, time_axis = calc_third_octave_levels(signal, fs)
+    # elif is_stationary == False:
+    #     spec_third, third_axis, time_axis = calc_third_octave_levels(signal, fs)
     np.squeeze(spec_third)
 
     output = {
